fix: AlarmResponseDto에 닉네임, 프로필 이미지 추가
