+++ friends_server/src/main/kotlin/com/friends/alarm/AlarmDtos.kt
@@ -1,6 +1,5 @@
 package com.friends.alarm
 
-import com.friends.alarm.entity.Alarm
 import com.friends.alarm.entity.AlarmType
 import java.time.LocalDateTime
 
@@ -9,18 +8,9 @@ data class AlarmResponseDto(
     val type: AlarmType,
     val message: String,
     val senderId: Long,
+    val senderProfileImage: String,
+    val nickname: String,
     val receiverId: Long,
     val invitedChatRoomId: Long? = null,
     val createdAt: LocalDateTime,
 )
-
-fun toAlarmResponseDto(alarm: Alarm) =
-    AlarmResponseDto(
-        id = alarm.id,
-        type = alarm.getType(),
-        message = alarm.message,
-        senderId = alarm.sender.id,
-        receiverId = alarm.receiver.id,
-        invitedChatRoomId = alarm.invitedChatRoom?.id,
-        createdAt = alarm.createdAt,
-    )
